Move SessionStore to common
